requirements.txt
Django==1.9.5
django-datetime-widget==0.9.3
django-npm==1.0.0
